DAY9-LangChain-Memory-Tools/WebSearch/requirements.txt
langchain 
langchain-google-genai
langchain-community
langchain-google-community
google-api-python-client
python-dotenv